# specs/001-invisiguard-core/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: InvisiGuard API
  description: API for invisible watermarking and geometric correction.
  version: 1.0.0
paths:
  /api/v1/embed:
    post:
      summary: Embed watermark into an image
      operationId: embedWatermark
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The source image (JPEG/PNG)
                text:
                  type: string
                  maxLength: 32
                  description: Text to embed
                alpha:
                  type: number
                  format: float
                  default: 1.0
                  minimum: 0.1
                  maximum: 5.0
                  description: Embedding strength
              required:
                - file
                - text
      responses:
        '200':
          description: Successful embedding
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      image_url:
                        type: string
                      psnr:
                        type: number
                        format: float
                      ssim:
                        type: number
                        format: float
        '400':
          description: Invalid input
        '500':
          description: Server error

  /api/v1/extract:
    post:
      summary: Extract watermark from a suspect image
      operationId: extractWatermark
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                original_file:
                  type: string
                  format: binary
                  description: Original image for alignment
                suspect_file:
                  type: string
                  format: binary
                  description: Suspect image (screenshot/photo)
              required:
                - original_file
                - suspect_file
      responses:
        '200':
          description: Successful extraction
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      decoded_text:
                        type: string
                      confidence:
                        type: number
                        format: float
                      is_match:
                        type: boolean
                      debug_info:
                        type: object
                        properties:
                          aligned_image_url:
                            type: string
                          matches_found:
                            type: integer
        '400':
          description: Invalid input
        '500':
          description: Server error